+++ all-is-cubes/src/math/aab.rs
@@ -4,7 +4,8 @@ use std::iter::FusedIterator;
 
 use cgmath::{EuclideanSpace as _, Point3, Vector3, Zero as _};
 
-use crate::math::{Face6, FreeCoordinate, Geometry, GridAab, GridCoordinate, GridPoint, Rgba};
+use crate::math::{Face6, FreeCoordinate, Geometry, GridAab, GridCoordinate, GridPoint};
+use crate::mesh::LineVertex;
 
 /// Axis-Aligned Box data type.
 ///
@@ -321,9 +322,9 @@ impl Geometry for Aab {
 
     fn wireframe_points<E>(&self, output: &mut E)
     where
-        E: Extend<(Point3<FreeCoordinate>, Option<Rgba>)>,
+        E: Extend<LineVertex>,
     {
-        let mut vertices = [(Point3::origin(), None); 24];
+        let mut vertices = [LineVertex::from(Point3::origin()); 24];
         let l = self.lower_bounds_p();
         let u = self.upper_bounds_p();
         for axis_0 in 0..3_usize {
@@ -332,20 +333,20 @@ impl Geometry for Aab {
             let axis_2 = (axis_0 + 2).rem_euclid(3);
             let mut p = l;
             // Walk from lower to upper in a helix.
-            vertices[vbase].0 = p;
+            vertices[vbase].position = p;
             p[axis_0] = u[axis_0];
-            vertices[vbase + 1].0 = p;
-            vertices[vbase + 2].0 = p;
+            vertices[vbase + 1].position = p;
+            vertices[vbase + 2].position = p;
             p[axis_1] = u[axis_1];
-            vertices[vbase + 3].0 = p;
-            vertices[vbase + 4].0 = p;
+            vertices[vbase + 3].position = p;
+            vertices[vbase + 4].position = p;
             p[axis_2] = u[axis_2];
-            vertices[vbase + 5].0 = p;
+            vertices[vbase + 5].position = p;
             // Go back and fill in the remaining bar.
             p[axis_2] = l[axis_2];
-            vertices[vbase + 6].0 = p;
+            vertices[vbase + 6].position = p;
             p[axis_0] = l[axis_0];
-            vertices[vbase + 7].0 = p;
+            vertices[vbase + 7].position = p;
         }
         output.extend(vertices);
     }
@@ -400,13 +401,13 @@ mod tests {
     #[test]
     fn aab_wireframe_smoke_test() {
         let aab = Aab::from_cube(Point3::new(1, 2, 3));
-        let mut wireframe: Vec<(Point3<FreeCoordinate>, Option<Rgba>)> = Vec::new();
+        let mut wireframe: Vec<LineVertex> = Vec::new();
         aab.wireframe_points(&mut wireframe);
-        for (vertex, color) in wireframe {
+        for LineVertex { position, color } in wireframe {
             assert!(color.is_none());
-            assert!(vertex.x == 1.0 || vertex.x == 2.0);
-            assert!(vertex.y == 2.0 || vertex.y == 3.0);
-            assert!(vertex.z == 3.0 || vertex.z == 4.0);
+            assert!(position.x == 1.0 || position.x == 2.0);
+            assert!(position.y == 2.0 || position.y == 3.0);
+            assert!(position.z == 3.0 || position.z == 4.0);
         }
     }